+++ tomography_preprocessing/tilt_series_alignment/__init__.py
@@ -1,5 +1,7 @@
-from .imod_fiducials import relion_batch_imod_fiducials
-from .imod_patch_tracking import align_tilt_series_in_imod_using_patch_tracking
-from .imod_generate_matrices import generate_imod_matrices
-from .aretomo import relion_batch_aretomo
+# Import decorated batch alignment functions first
+from .aretomo import batch_aretomo as _batch_aretomo
+from .imod import batch_fiducials as _batch_imod_fiducials
+from .imod import batch_patch_tracking as _batch_imod_patch_tracking
+
+# Then import the cli, it will be decorated with all programs
 from ._cli import cli
