+++ tomography_preprocessing/tilt_series_alignment/imod/generate_matrices.py
@@ -1,4 +1,3 @@
-import os
 from pathlib import Path
 from typing import Tuple
 
@@ -6,9 +5,9 @@ import numpy as np
 import starfile
 import typer
 
-import tomography_preprocessing.tilt_series_alignment._utils
-from ._cli import cli
-from .. import utils
+from .._cli import cli
+from .._job_utils import create_alignment_job_directory_structure
+from ... import utils
 
 
 @cli.command(name='IMOD:generate-matrices')
@@ -20,21 +19,22 @@ def generate_imod_matrices(
     tilt_series_metadata = utils.star.iterate_tilt_series_metadata(
         tilt_series_star_file=tilt_series_star_file,
     )
-    tilt_series_directory, imod_alignments_directory = \
-        tomography_preprocessing.tilt_series_alignment._utils.create_alignment_job_directory_structure(output_directory)
+    image_directory, alignments_directory = \
+        create_alignment_job_directory_structure(output_directory)
     tilt_image_star_files = []
 
     for tilt_series_id, tilt_series_df, tilt_image_df in tilt_series_metadata:
         euler_angles = tilt_image_df[['rlnAngleRot', 'rlnAngleTilt', 'rlnAnglePsi']]
         shifts = tilt_image_df[['rlnOriginXAngst', 'rlnOriginYAngst']]
         shifts /= tilt_series_df['rlnMicrographOriginalPixelSize']
-        matrices = utils.imod.relion_tilt_series_alignment_parameters_to_relion_matrix(
+        matrices = tomography_preprocessing.tilt_series_alignment.imod.imod.relion_tilt_series_alignment_parameters_to_relion_matrix(
             shifts=shifts,
             euler_angles=euler_angles,
-            tilt_image_dimensions=utils.mrc.get_image_dimensions(tilt_image_df['rlnMicrographName'][0])[:2],
+            tilt_image_dimensions=utils.mrc.get_image_dimensions(
+                tilt_image_df['rlnMicrographName'][0])[:2],
             tomogram_dimensions=np.array(tomogram_dimensions)
         )
-        tilt_image_star_file = tilt_series_directory / f'{tilt_series_id}.star'
+        tilt_image_star_file = image_directory / f'{tilt_series_id}.star'
         tilt_image_star_files.append(tilt_image_star_file)
         projection_matrix_labels = [f'rlnTomoProj{ax}' for ax in 'XYZW']
         for idx, label in enumerate(projection_matrix_labels):
@@ -48,7 +48,3 @@ def generate_imod_matrices(
     star = starfile.read(tilt_series_star_file, always_dict=True)
     star['global']['rlnTomoTiltSeriesStarFile'] = tilt_image_star_files
     starfile.write(star, output_directory / 'aligned_tilt_series_with_matrices.star')
-
-
-
-
